Trait token_migration::Trait

pub trait Trait: Config {
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type BlockNumberToBalance: Convert<Self::BlockNumber, <Self::Currency as Currency<Self::AccountId>>::Balance>;
    type VestingMilestones: Get<u8>;
    type VestingDuration: Get<u32>;
}

The pallet’s configuration trait.

Associated Types

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type Currency: ReservableCurrency<Self::AccountId>

Currency type that supports locking

type BlockNumberToBalance: Convert<Self::BlockNumber, <Self::Currency as Currency<Self::AccountId>>::Balance>

Convert the block number into a balance.

type VestingMilestones: Get<u8>

NOTE: Both VestingMilestones and VestingDuration must be > 0. The ideal situation would be to have a compile time assertion in this pallet but since the static assertions pallet does not work with generics (https://github.com/nvzqz/static-assertions-rs/issues/21), the check has been moved to the runtime instantiation. If this pallet is used outside this project, corresponding checks should be done. Vesting happens in milestones. The total duration is sub-divided into equal duration milestones and for each milestone, proportional balance is vested. This might be moved to a storage item if needs to be configurable but is less likely.

type VestingDuration: Get<u32>

Vesting duration in number of blocks.

Loading content...