Crate token_migration

Structs

Module
The module declaration.

OnlyMigrator
Signed extension to ensure that only a Migrator can send the migrate extrinsic. This is necessary to prevent a migrate call done by non-Migrator to even enter a block.

Enums

Call
Dispatchable calls.

Error
Errors for the module.

RawEvent
Events for this module.

Traits

Trait
The pallet's configuration trait.

Type Definitions

Event
RawEvent specialized for the configuration Trait