Crate token_migration

Pallet for token migration from ERC-20 to Dock’s native token. Migrators are assumed to not be adversarial and not do DoS attacks on the chain and are the exchanges in practice. Migrators are given balance by Dock to fulfil expected migration requests and Dock needs to ensure that migrators do hold the correct amount of ERC-20 and that they will lock it (request signature from addresses to check ownership and monitor addresses or ask them to transfer to the Vault). Also bonus amounts are calculated post the initial migration and depend on the participating holder and their vesting choice which Dock should verify with external migrators.

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

Bonus
Struct to encode all the bonuses of an account.

Bonuses
Tracks swap and vesting bonuses that will be given to holders

Module
The module declaration.

OnlyMigrator
Signed extension to ensure that only a Migrator can send the migrate extrinsic. This is necessary to prevent a migrate call done by non-Migrator to even enter a block.

Enums

Call
Dispatchable calls.

Error
Errors for the module.

RawEvent
Events for this module.

Traits

Trait
The pallet’s configuration trait.

Type Definitions

Event
RawEvent specialized for the configuration Config